+++ ch6/exercises/test-randomlyPermuteArray.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+g++ randomlyPermuteArray.cpp -o randomlyPermuteArray
+./randomlyPermuteArray
+
+rm randomlyPermuteArray