+++ ch6/exercises/test-reverseArray.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+g++ reverseArray.cpp -o reverseArray
+./reverseArray
+
+rm reverseArray